+++ radio_interferometer/__init__.py
@@ -1,3 +1,3 @@
 """Radio interferometry FX correlator application."""
 
-__version__ = "0.5.3-iota"
+__version__ = "0.6.0-mu"